Refactor json serialization code
* Renamed encodePrimitiveData to encodeValue
* Changed return type of toJsonWithMetadata to not expose JsonObject in publec API
* Moved internal and private functions from json.kt to new writeJson.kt / readJson.kt files under impl.io package
