+++ core/src/main/kotlin/org/jetbrains/kotlinx/dataframe/io/json.kt
@@ -1,74 +1,28 @@
 package org.jetbrains.kotlinx.dataframe.io
 
-import com.beust.klaxon.JsonArray
-import com.beust.klaxon.JsonObject
-import com.beust.klaxon.KlaxonJson
 import com.beust.klaxon.Parser
 import com.beust.klaxon.json
-import org.jetbrains.kotlinx.dataframe.AnyCol
 import org.jetbrains.kotlinx.dataframe.AnyFrame
 import org.jetbrains.kotlinx.dataframe.AnyRow
-import org.jetbrains.kotlinx.dataframe.ColumnsContainer
-import org.jetbrains.kotlinx.dataframe.DataColumn
 import org.jetbrains.kotlinx.dataframe.DataFrame
 import org.jetbrains.kotlinx.dataframe.DataRow
 import org.jetbrains.kotlinx.dataframe.api.JsonPath
 import org.jetbrains.kotlinx.dataframe.api.KeyValueProperty
-import org.jetbrains.kotlinx.dataframe.api.cast
-import org.jetbrains.kotlinx.dataframe.api.columnOf
-import org.jetbrains.kotlinx.dataframe.api.concat
-import org.jetbrains.kotlinx.dataframe.api.dataFrameOf
-import org.jetbrains.kotlinx.dataframe.api.firstOrNull
-import org.jetbrains.kotlinx.dataframe.api.getColumn
-import org.jetbrains.kotlinx.dataframe.api.indices
-import org.jetbrains.kotlinx.dataframe.api.isList
-import org.jetbrains.kotlinx.dataframe.api.mapIndexed
-import org.jetbrains.kotlinx.dataframe.api.name
-import org.jetbrains.kotlinx.dataframe.api.named
-import org.jetbrains.kotlinx.dataframe.api.rows
-import org.jetbrains.kotlinx.dataframe.api.schema
 import org.jetbrains.kotlinx.dataframe.api.single
-import org.jetbrains.kotlinx.dataframe.api.splitInto
-import org.jetbrains.kotlinx.dataframe.api.toDataFrame
 import org.jetbrains.kotlinx.dataframe.codeGen.DefaultReadDfMethod
 import org.jetbrains.kotlinx.dataframe.codeGen.DefaultReadJsonMethod
 import org.jetbrains.kotlinx.dataframe.columns.ColumnGroup
-import org.jetbrains.kotlinx.dataframe.columns.ColumnKind
 import org.jetbrains.kotlinx.dataframe.columns.FrameColumn
-import org.jetbrains.kotlinx.dataframe.impl.ColumnNameGenerator
-import org.jetbrains.kotlinx.dataframe.impl.DataCollectorBase
-import org.jetbrains.kotlinx.dataframe.impl.asList
-import org.jetbrains.kotlinx.dataframe.impl.columns.createColumn
-import org.jetbrains.kotlinx.dataframe.impl.commonType
-import org.jetbrains.kotlinx.dataframe.impl.createDataCollector
-import org.jetbrains.kotlinx.dataframe.impl.guessValueType
-import org.jetbrains.kotlinx.dataframe.impl.schema.DataFrameSchemaImpl
-import org.jetbrains.kotlinx.dataframe.impl.schema.extractSchema
-import org.jetbrains.kotlinx.dataframe.impl.schema.intersectSchemas
-import org.jetbrains.kotlinx.dataframe.impl.splitByIndices
+import org.jetbrains.kotlinx.dataframe.impl.io.encodeDataFrameWithMetadata
+import org.jetbrains.kotlinx.dataframe.impl.io.encodeFrame
+import org.jetbrains.kotlinx.dataframe.impl.io.encodeRow
+import org.jetbrains.kotlinx.dataframe.impl.io.readJson
 import org.jetbrains.kotlinx.dataframe.io.JSON.TypeClashTactic
 import org.jetbrains.kotlinx.dataframe.io.JSON.TypeClashTactic.ANY_COLUMNS
 import org.jetbrains.kotlinx.dataframe.io.JSON.TypeClashTactic.ARRAY_AND_VALUE_COLUMNS
-import org.jetbrains.kotlinx.dataframe.io.SerializationKeys.COLUMNS
-import org.jetbrains.kotlinx.dataframe.io.SerializationKeys.DATA
-import org.jetbrains.kotlinx.dataframe.io.SerializationKeys.KIND
-import org.jetbrains.kotlinx.dataframe.io.SerializationKeys.KOTLIN_DATAFRAME
-import org.jetbrains.kotlinx.dataframe.io.SerializationKeys.METADATA
-import org.jetbrains.kotlinx.dataframe.io.SerializationKeys.NCOL
-import org.jetbrains.kotlinx.dataframe.io.SerializationKeys.NROW
-import org.jetbrains.kotlinx.dataframe.io.SerializationKeys.VERSION
-import org.jetbrains.kotlinx.dataframe.ncol
-import org.jetbrains.kotlinx.dataframe.nrow
-import org.jetbrains.kotlinx.dataframe.schema.ColumnSchema
-import org.jetbrains.kotlinx.dataframe.type
-import org.jetbrains.kotlinx.dataframe.typeClass
-import org.jetbrains.kotlinx.dataframe.values
 import java.io.File
 import java.io.InputStream
 import java.net.URL
-import kotlin.reflect.KType
-import kotlin.reflect.KTypeProjection
-import kotlin.reflect.full.createType
 import kotlin.reflect.typeOf
 
 public class JSON(
@@ -152,6 +106,9 @@ public class JSON(
         ANY_COLUMNS,
     }
 }
+
+public const val arrayColumnName: String = "array"
+public const val valueColumnName: String = "value"
 
 /**
  * @param file Where to fetch the Json as [InputStream] to be converted to a [DataFrame].
@@ -303,752 +260,34 @@ public fun DataRow.Companion.readJsonStr(
     typeClashTactic: TypeClashTactic = ARRAY_AND_VALUE_COLUMNS,
 ): AnyRow = DataFrame.readJsonStr(text, header, keyValuePaths, typeClashTactic).single()
 
-private fun readJson(
-    parsed: Any?,
-    header: List<String>,
-    keyValuePaths: List<JsonPath> = emptyList(),
-    typeClashTactic: TypeClashTactic = ARRAY_AND_VALUE_COLUMNS,
-): DataFrame<*> {
-    val df: AnyFrame = when (typeClashTactic) {
-        ARRAY_AND_VALUE_COLUMNS -> {
-            when (parsed) {
-                is JsonArray<*> -> fromJsonListArrayAndValueColumns(
-                    records = parsed.value,
-                    header = header,
-                    keyValuePaths = keyValuePaths,
-                )
-
-                else -> fromJsonListArrayAndValueColumns(
-                    records = listOf(parsed),
-                    keyValuePaths = keyValuePaths,
-                )
-            }
-        }
-
-        ANY_COLUMNS -> {
-            when (parsed) {
-                is JsonArray<*> -> fromJsonListAnyColumns(
-                    records = parsed.value,
-                    header = header,
-                    keyValuePaths = keyValuePaths,
-                )
-
-                else -> fromJsonListAnyColumns(
-                    records = listOf(parsed),
-                    keyValuePaths = keyValuePaths,
-                )
-            }
-        }
-    }
-    return df.unwrapUnnamedColumns()
-}
-
-private fun DataFrame<Any?>.unwrapUnnamedColumns() =
-    dataFrameOf(columns().map { it.unwrapUnnamedColumn() })
-
-private fun AnyCol.unwrapUnnamedColumn() = if (this is UnnamedColumn) col else this
-
-private enum class AnyColType {
-    ANY,
-    ARRAYS,
-    OBJECTS,
-}
-
-internal interface AnyKeyValueProperty : KeyValueProperty<Any?> {
-    override val value: Any?
-}
-
-/**
- * Json to DataFrame converter that creates [Any] columns.
- * A.k.a. [TypeClashTactic.ANY_COLUMNS].
- *
- * @param records List of json elements to be converted to a [DataFrame].
- * @param keyValuePaths List of [JsonPath]s where instead of a [ColumnGroup], a [FrameColumn]<[KeyValueProperty]>
- *     will be created.
- * @param header Optional list of column names. If given, [records] will be read like an object with [header] being the keys.
- * @return [DataFrame] from the given [records].
- */
-internal fun fromJsonListAnyColumns(
-    records: List<*>,
-    keyValuePaths: List<JsonPath> = emptyList(),
-    header: List<String> = emptyList(),
-    jsonPath: JsonPath = JsonPath(),
-): AnyFrame {
-    var hasPrimitive = false
-    var hasArray = false
-    var hasObject = false
-
-    // list element type can be JsonObject, JsonArray or primitive
-    val nameGenerator = ColumnNameGenerator()
-    records.forEach {
-        when (it) {
-            is JsonObject -> {
-                hasObject = true
-                it.entries.forEach {
-                    nameGenerator.addIfAbsent(it.key)
-                }
-            }
-
-            is JsonArray<*> -> hasArray = true
-            null -> Unit
-            else -> hasPrimitive = true
-        }
-    }
-
-    val colType = when {
-        hasArray && !hasPrimitive && !hasObject -> AnyColType.ARRAYS
-        hasObject && !hasPrimitive && !hasArray -> AnyColType.OBJECTS
-        else -> AnyColType.ANY
-    }
-    val justPrimitives = hasPrimitive && !hasArray && !hasObject
-    val isKeyValue = keyValuePaths.any { jsonPath.matches(it) }
-
-    if (isKeyValue && colType != AnyColType.OBJECTS) {
-        error("Key value path $jsonPath does not match objects.")
-    }
-
-    @Suppress("KotlinConstantConditions")
-    val columns: List<AnyCol> = when {
-        // Create one column of type Any? (or guessed primitive type) from all the records
-        colType == AnyColType.ANY -> {
-            val collector: DataCollectorBase<Any?> =
-                if (justPrimitives) createDataCollector(records.size) // guess the type
-                else createDataCollector(records.size, typeOf<Any?>()) // use Any?
-
-            val nanIndices = mutableListOf<Int>()
-            records.forEachIndexed { i, v ->
-                when (v) {
-                    is JsonObject -> {
-                        val parsed =
-                            fromJsonListAnyColumns(
-                                records = listOf(v),
-                                keyValuePaths = keyValuePaths,
-                                jsonPath = jsonPath.replaceLastWildcardWithIndex(i),
-                            )
-                        collector.add(
-                            if (parsed.isSingleUnnamedColumn()) (parsed.getColumn(0) as UnnamedColumn).col.values.first()
-                            else parsed.firstOrNull() ?: DataRow.empty
-                        )
-                    }
-
-                    is JsonArray<*> -> {
-                        val parsed = fromJsonListAnyColumns(
-                            records = v,
-                            keyValuePaths = keyValuePaths,
-                            jsonPath = jsonPath.replaceLastWildcardWithIndex(i).appendArrayWithWildcard(),
-                        )
-                        collector.add(
-                            if (parsed.isSingleUnnamedColumn()) (parsed.getColumn(0) as UnnamedColumn).col.values.asList()
-                            else parsed.unwrapUnnamedColumns()
-                        )
-                    }
-
-                    "NaN" -> {
-                        nanIndices.add(i)
-                        collector.add(null)
-                    }
-
-                    else -> collector.add(v)
-                }
-            }
-            val column = collector.toColumn(valueColumnName)
-            val res = if (nanIndices.isNotEmpty()) {
-                fun <C> DataColumn<C>.updateNaNs(nanValue: C): DataColumn<C> {
-                    var j = 0
-                    var nextNanIndex = nanIndices[j]
-                    return mapIndexed(column.type) { i, v ->
-                        if (i == nextNanIndex) {
-                            j++
-                            nextNanIndex = if (j < nanIndices.size) nanIndices[j] else -1
-                            nanValue
-                        } else v
-                    }
-                }
-                when (column.typeClass) {
-                    Double::class -> column.cast<Double?>().updateNaNs(Double.NaN)
-                    Float::class -> column.cast<Float?>().updateNaNs(Float.NaN)
-                    String::class -> column.cast<String?>().updateNaNs("NaN")
-                    else -> column
-                }
-            } else column
-            listOf(UnnamedColumn(res))
-        }
-
-        // Create one column of type FrameColumn, or List<> from all the records if they are all arrays
-        colType == AnyColType.ARRAYS -> {
-            val values = mutableListOf<Any?>()
-            val startIndices = ArrayList<Int>()
-            records.forEach {
-                startIndices.add(values.size)
-                when (it) {
-                    is JsonArray<*> -> values.addAll(it.value)
-                    null -> Unit
-                    else -> error("Expected JsonArray, got $it")
-                }
-            }
-            val parsed = fromJsonListAnyColumns(
-                records = values,
-                keyValuePaths = keyValuePaths,
-                jsonPath = jsonPath.appendArrayWithWildcard(),
-            )
-
-            val res = when {
-                parsed.isSingleUnnamedColumn() -> {
-                    val col = (parsed.getColumn(0) as UnnamedColumn).col
-                    val elementType = col.type
-                    val values = col.values.asList().splitByIndices(startIndices.asSequence()).toList()
-                    DataColumn.createValueColumn(
-                        name = arrayColumnName,
-                        values = values,
-                        type = List::class.createType(listOf(KTypeProjection.invariant(elementType))),
-                    )
-                }
-
-                else -> DataColumn.createFrameColumn(
-                    name = arrayColumnName, // will be erased
-                    df = parsed.unwrapUnnamedColumns(),
-                    startIndices = startIndices,
-                )
-            }
-            listOf(UnnamedColumn(res))
-        }
-
-        // Create one column of type FrameColumn<KeyValueProperty>
-        colType == AnyColType.OBJECTS && isKeyValue -> {
-            // collect the value types to make sure Value columns with lists and other values aren't all turned into lists
-            val valueTypes = mutableSetOf<KType>()
-            val dataFrames = records.map {
-                when (it) {
-                    is JsonObject -> {
-                        val map = it.map.mapValues { (key, value) ->
-                            val parsed = fromJsonListAnyColumns(
-                                records = listOf(value),
-                                keyValuePaths = keyValuePaths,
-                                jsonPath = jsonPath.append(key),
-                            )
-                            if (parsed.isSingleUnnamedColumn()) (parsed.getColumn(0) as UnnamedColumn).col.values.first()
-                            else parsed.unwrapUnnamedColumns().firstOrNull()
-                        }
-                        val valueType = map.values.map {
-                            guessValueType(sequenceOf(it))
-                        }.commonType()
-
-                        valueTypes += valueType
-
-                        dataFrameOf(
-                            columnOf(*map.keys.toTypedArray()).named(KeyValueProperty<*>::key.name),
-                            createColumn(values = map.values, suggestedType = valueType, guessType = false)
-                                .named(KeyValueProperty<*>::value.name),
-                        )
-                    }
-
-                    null -> DataFrame.emptyOf<AnyKeyValueProperty>()
-                    else -> error("Expected JsonObject, got $it")
-                }
-            }
-
-            val valueColumns = dataFrames.map { it[KeyValueProperty<*>::value.name] }
-            val valueColumnSchema = when {
-                // in these cases we can safely combine the columns to get a single column schema
-                valueColumns.all { it is ColumnGroup<*> } || valueColumns.all { it is FrameColumn<*> } ->
-                    valueColumns.concat().extractSchema()
-                // to avoid listification, we create the value columns schema ourselves (https://github.com/Kotlin/dataframe/issues/184)
-                else -> ColumnSchema.Value(valueTypes.commonType())
-            }
-
-            listOf(
-                UnnamedColumn(
-                    DataColumn.createFrameColumn(
-                        name = valueColumnName, // will be erased unless at top-level
-                        groups = dataFrames,
-                        schema = lazy {
-                            DataFrameSchemaImpl(
-                                columns = mapOf(
-                                    KeyValueProperty<*>::key.name to ColumnSchema.Value(typeOf<String>()),
-                                    KeyValueProperty<*>::value.name to valueColumnSchema,
-                                )
-                            )
-                        },
-                    )
-                )
-            )
-        }
-
-        // Create multiple columns from all the records if they are all objects, merging the objects in essence
-        colType == AnyColType.OBJECTS && !isKeyValue -> {
-            nameGenerator.names.map { colName ->
-                val values = ArrayList<Any?>(records.size)
-
-                records.forEach {
-                    when (it) {
-                        is JsonObject -> values.add(it[colName])
-                        null -> values.add(null)
-                        else -> error("Expected JsonObject, got $it")
-                    }
-                }
-
-                val parsed = fromJsonListAnyColumns(
-                    records = values,
-                    keyValuePaths = keyValuePaths,
-                    jsonPath = jsonPath.append(colName),
-                )
-                when {
-                    parsed.ncol == 0 ->
-                        DataColumn.createValueColumn(
-                            name = colName,
-                            values = arrayOfNulls<Any?>(values.size).toList(),
-                            type = typeOf<Any?>(),
-                        )
-
-                    parsed.isSingleUnnamedColumn() ->
-                        (parsed.getColumn(0) as UnnamedColumn).col.rename(colName)
-
-                    else ->
-                        DataColumn.createColumnGroup(colName, parsed.unwrapUnnamedColumns()) as AnyCol
-                }
-            }
-        }
-
-        else -> error("")
-    }
-
-    return when {
-        columns.isEmpty() -> DataFrame.empty(records.size)
-
-        columns.size == 1 && hasArray && header.isNotEmpty() && columns[0].typeClass == List::class ->
-            columns[0]
-                .cast<List<*>>()
-                .splitInto(*header.toTypedArray())
-
-        else -> columns.toDataFrame()
-    }
-}
-
-public const val arrayColumnName: String = "array"
-public const val valueColumnName: String = "value"
-
-private fun AnyFrame.isSingleUnnamedColumn() = ncol == 1 && getColumn(0) is UnnamedColumn
-
-/**
- * Json to DataFrame converter that creates allows creates `value` and `array` accessors
- * instead of [Any] columns.
- * A.k.a. [TypeClashTactic.ARRAY_AND_VALUE_COLUMNS].
- *
- * @param records List of json elements to be converted to a [DataFrame].
- * @param keyValuePaths List of [JsonPath]s where instead of a [ColumnGroup], a [FrameColumn]<[KeyValueProperty]>
- *     will be created.
- * @param header Optional list of column names. If given, [records] will be read like an object with [header] being the keys.
- * @return [DataFrame] from the given [records].
- */
-internal fun fromJsonListArrayAndValueColumns(
-    records: List<*>,
-    keyValuePaths: List<JsonPath> = emptyList(),
-    header: List<String> = emptyList(),
-    jsonPath: JsonPath = JsonPath(),
-): AnyFrame {
-    var hasPrimitive = false
-    var hasArray = false
-    val isKeyValue = keyValuePaths.any { jsonPath.matches(it) }
-
-    // list element type can be JsonObject, JsonArray or primitive
-    // So first, we gather all properties of objects to merge including "array" and "value" if needed
-    // so the resulting type of a property with instances 123, ["abc"], and { "a": 1, "b": 2 } will be
-    // { array: List<String>, value: Int?, a: Int?, b: Int? }
-    // and instances will look like
-    // { "array": [], "value": 123, "a": null, "b": null }
-
-    val nameGenerator = ColumnNameGenerator()
-    records.forEach {
-        when (it) {
-            is JsonObject -> it.entries.forEach {
-                nameGenerator.addIfAbsent(it.key)
-            }
-
-            is JsonArray<*> -> hasArray = true
-            null -> Unit
-            else -> hasPrimitive = true
-        }
-    }
-    if (records.all { it == null }) hasPrimitive = true
-
-    // Add a value column to the collected names if needed
-    val valueColumn = if (hasPrimitive || records.isEmpty()) {
-        nameGenerator.addUnique(valueColumnName)
-    } else null
-
-    // Add an array column to the collected names if needed
-    val arrayColumn = if (hasArray) {
-        nameGenerator.addUnique(arrayColumnName)
-    } else null
-
-    // only properties that consist of just objects (or are empty) can be merged to key/value FrameColumns
-    if (isKeyValue && (hasPrimitive || hasArray)) {
-        error("Key value path $jsonPath does not match objects.")
-    }
-
-    // Create columns from the collected names
-    val columns: List<AnyCol> = when {
-        // instead of using the names, generate a single key/value frame column
-        isKeyValue -> {
-            val dataFrames = records.map {
-                when (it) {
-                    is JsonObject -> {
-                        val map = it.map.mapValues { (key, value) ->
-                            val parsed = fromJsonListArrayAndValueColumns(
-                                records = listOf(value),
-                                keyValuePaths = keyValuePaths,
-                                jsonPath = jsonPath.append(key),
-                            )
-                            if (parsed.isSingleUnnamedColumn()) (parsed.getColumn(0) as UnnamedColumn).col.values.first()
-                            else parsed.unwrapUnnamedColumns().firstOrNull()
-                        }
-                        val valueType =
-                            map.values.map { guessValueType(sequenceOf(it)) }
-                                .commonType()
-
-                        dataFrameOf(
-                            columnOf(*map.keys.toTypedArray()).named(KeyValueProperty<*>::key.name),
-                            createColumn(
-                                values = map.values,
-                                suggestedType = valueType,
-                                guessType = false,
-                            ).named(KeyValueProperty<*>::value.name),
-                        )
-                    }
-
-                    null -> DataFrame.emptyOf<AnyKeyValueProperty>()
-                    else -> error("Expected JsonObject, got $it")
-                }
-            }
-
-            listOf(
-                UnnamedColumn(
-                    DataColumn.createFrameColumn(
-                        name = valueColumnName, // will be erased unless at top-level
-                        groups = dataFrames,
-                        schema = lazy {
-                            dataFrames.mapNotNull { it.takeIf { it.nrow > 0 }?.schema() }.intersectSchemas()
-                        },
-                    )
-                )
-            )
-        }
-
-        // generate columns using the collected names
-        else ->
-            nameGenerator.names.map { colName ->
-                when {
-                    // Collect primitive values from records into the `value` column if needed
-                    colName == valueColumn && (hasPrimitive || records.isEmpty()) -> {
-                        val collector = createDataCollector(records.size)
-                        val nanIndices = mutableListOf<Int>()
-                        records.forEachIndexed { i, v ->
-                            when (v) {
-                                is JsonObject -> collector.add(null)
-                                is JsonArray<*> -> collector.add(null)
-                                "NaN" -> {
-                                    nanIndices.add(i)
-                                    collector.add(null)
-                                }
-
-                                else -> collector.add(v)
-                            }
-                        }
-                        val column = collector.toColumn(colName)
-                        val res = if (nanIndices.isNotEmpty()) {
-                            fun <C> DataColumn<C>.updateNaNs(nanValue: C): DataColumn<C> {
-                                var j = 0
-                                var nextNanIndex = nanIndices[j]
-                                return mapIndexed(column.type) { i, v ->
-                                    if (i == nextNanIndex) {
-                                        j++
-                                        nextNanIndex = if (j < nanIndices.size) nanIndices[j] else -1
-                                        nanValue
-                                    } else v
-                                }
-                            }
-                            when (column.typeClass) {
-                                Double::class -> column.cast<Double?>().updateNaNs(Double.NaN)
-                                Float::class -> column.cast<Float?>().updateNaNs(Float.NaN)
-                                String::class -> column.cast<String?>().updateNaNs("NaN")
-                                else -> column
-                            }
-                        } else column
-                        UnnamedColumn(res)
-                    }
-
-                    // Collect arrays from records into the `array` column if needed
-                    colName == arrayColumn && hasArray -> {
-                        val values = mutableListOf<Any?>()
-                        val startIndices = ArrayList<Int>()
-                        records.forEach {
-                            startIndices.add(values.size)
-                            if (it is JsonArray<*>) values.addAll(it.value)
-                        }
-                        val parsed = fromJsonListArrayAndValueColumns(
-                            records = values,
-                            keyValuePaths = keyValuePaths,
-                            jsonPath = jsonPath.appendArrayWithWildcard(),
-                        )
-
-                        val res = when {
-                            parsed.isSingleUnnamedColumn() -> {
-                                val col = (parsed.getColumn(0) as UnnamedColumn).col
-                                val elementType = col.type
-                                val values = col.values.asList().splitByIndices(startIndices.asSequence()).toList()
-                                DataColumn.createValueColumn(
-                                    name = colName,
-                                    values = values,
-                                    type = List::class.createType(listOf(KTypeProjection.invariant(elementType))),
-                                )
-                            }
-
-                            else -> DataColumn.createFrameColumn(colName, parsed.unwrapUnnamedColumns(), startIndices)
-                        }
-                        UnnamedColumn(res)
-                    }
-
-                    // Collect the current column name as property from the objects in records
-                    else -> {
-                        val values = ArrayList<Any?>(records.size)
-                        records.forEach {
-                            when (it) {
-                                is JsonObject -> values.add(it[colName])
-                                else -> values.add(null)
-                            }
-                        }
-
-                        val parsed = fromJsonListArrayAndValueColumns(
-                            records = values,
-                            keyValuePaths = keyValuePaths,
-                            jsonPath = jsonPath.append(colName),
-                        )
-                        when {
-                            parsed.ncol == 0 ->
-                                DataColumn.createValueColumn(
-                                    name = colName,
-                                    values = arrayOfNulls<Any?>(values.size).toList(),
-                                    type = typeOf<Any?>(),
-                                )
-
-                            parsed.isSingleUnnamedColumn() ->
-                                (parsed.getColumn(0) as UnnamedColumn).col.rename(colName)
-
-                            else ->
-                                DataColumn.createColumnGroup(colName, parsed.unwrapUnnamedColumns()) as AnyCol
-                        }
-                    }
-                }
-            }
-    }
-
-    return when {
-        columns.isEmpty() ->
-            DataFrame.empty(records.size)
-
-        columns.size == 1 && hasArray && header.isNotEmpty() && columns[0].typeClass == List::class ->
-            columns[0]
-                .cast<List<*>>()
-                .splitInto(*header.toTypedArray())
-
-        else ->
-            columns.toDataFrame()
-    }
-}
-
-// we need it to check if AnyFrame created by recursive call has single unnamed column,
-// unnamed column means this column is not created from field of a record [{"value": 1}, {"value": 2}],
-// but filtered values [1, { ... }, []] -> [1, null, null]
-// or arrays: [1, { ...}, []] -> [null, null, []]
-private class UnnamedColumn(val col: DataColumn<Any?>) : DataColumn<Any?> by col
-
-private val valueTypes =
-    setOf(Boolean::class, Double::class, Int::class, Float::class, Long::class, Short::class, Byte::class)
-
-internal fun KlaxonJson.encodeRow(frame: ColumnsContainer<*>, index: Int): JsonObject? {
-    val values = frame.columns().map { col ->
-        when (col) {
-            is ColumnGroup<*> -> encodeRow(col, index)
-            is FrameColumn<*> -> encodeFrame(col[index])
-            else -> encodePrimitiveData(col, index)
-        }.let { col.name to it }
-    }
-    if (values.isEmpty()) return null
-    return obj(values)
-}
-
-internal object SerializationKeys {
-    const val DATA = "data"
-    const val METADATA = "metadata"
-    const val KIND = "kind"
-    const val NCOL = "ncol"
-    const val NROW = "nrow"
-    const val VERSION = "\$version"
-    const val COLUMNS = "columns"
-    const val KOTLIN_DATAFRAME = "kotlin_dataframe"
-}
-
-internal const val SERIALIZATION_VERSION = "2.0.0"
-
-internal fun KlaxonJson.encodeRowWithMetadata(
-    frame: ColumnsContainer<*>,
-    index: Int,
-    rowLimit: Int? = null
-): JsonObject? {
-    val values = frame.columns().map { col ->
-        when (col) {
-            is ColumnGroup<*> -> obj(
-                DATA to encodeRowWithMetadata(col, index, rowLimit),
-                METADATA to obj(KIND to ColumnKind.Group.toString())
-            )
-
-            is FrameColumn<*> -> {
-                val data = if (rowLimit == null) encodeFrameWithMetadata(col[index])
-                else encodeFrameWithMetadata(col[index].rows().take(rowLimit).toDataFrame(), rowLimit)
-                obj(
-                    DATA to data,
-                    METADATA to obj(
-                        KIND to ColumnKind.Frame.toString(),
-                        NCOL to col[index].ncol,
-                        NROW to col[index].nrow
-                    )
-                )
-            }
-
-            else -> encodePrimitiveData(col, index)
-        }.let { col.name to it }
-    }
-    if (values.isEmpty()) return null
-    return obj(values)
-}
-
-internal fun KlaxonJson.encodePrimitiveData(col: AnyCol, index: Int): Any? = when {
-    col.isList() -> col[index]?.let { array(it as List<*>) } ?: array()
-    col.typeClass in valueTypes -> {
-        val v = col[index]
-        if ((v is Double && v.isNaN()) || (v is Float && v.isNaN())) {
-            v.toString()
-        } else v
-    }
-
-    else -> col[index]?.toString()
-}
-
-internal fun KlaxonJson.encodeFrameWithMetadata(frame: AnyFrame, rowLimit: Int? = null): JsonArray<*> {
-    val valueColumn = frame.extractValueColumn()
-    val arrayColumn = frame.extractArrayColumn()
-
-    val arraysAreFrames = arrayColumn?.kind() == ColumnKind.Frame
-
-    val data = frame.indices().map { rowIndex ->
-        valueColumn
-            ?.get(rowIndex)
-            ?: arrayColumn?.get(rowIndex)
-                ?.let {
-                    if (arraysAreFrames) encodeFrameWithMetadata(it as AnyFrame, rowLimit) else null
-                }
-            ?: encodeRowWithMetadata(frame, rowIndex, rowLimit)
-    }
-
-    return array(data)
-}
-
-internal fun AnyFrame.extractValueColumn(): DataColumn<*>? {
-    val allColumns = columns()
-
-    return allColumns.filter { it.name.startsWith(valueColumnName) }
-        .takeIf { isPossibleToFindUnnamedColumns }
-        ?.maxByOrNull { it.name }?.let { valueCol ->
-            if (valueCol.kind() != ColumnKind.Value) { // check that value in this column is not null only when other values are null
-                null
-            } else {
-                // check that value in this column is not null only when other values are null
-                val isValidValueColumn = rows().all { row ->
-                    if (valueCol[row] != null) {
-                        allColumns.all { col ->
-                            if (col.name != valueCol.name) col[row] == null
-                            else true
-                        }
-                    } else true
-                }
-                if (isValidValueColumn) valueCol
-                else null
-            }
-        }
-}
-
-// if there is only 1 column, then `isValidValueColumn` always true.
-// But at the same time, we shouldn't treat dataFrameOf("value")(1,2,3) like unnamed column
-// because it was created by user.
-internal val AnyFrame.isPossibleToFindUnnamedColumns: Boolean
-    get() = columns().size != 1
-
-internal fun AnyFrame.extractArrayColumn(): DataColumn<*>? {
-    val allColumns = columns()
-
-    return columns().filter { it.name.startsWith(arrayColumnName) }
-        .takeIf { isPossibleToFindUnnamedColumns }
-        ?.maxByOrNull { it.name }?.let { arrayCol ->
-            if (arrayCol.kind() == ColumnKind.Group) null
-            else {
-                // check that value in this column is not null only when other values are null
-                val isValidArrayColumn = rows().all { row ->
-                    if (arrayCol[row] != null) {
-                        allColumns.all { col ->
-                            if (col.name != arrayCol.name) col[row] == null
-                            else true
-                        }
-                    } else true
-                }
-                if (isValidArrayColumn) arrayCol
-                else null
-            }
-        }
-}
-
-internal fun KlaxonJson.encodeFrame(frame: AnyFrame): JsonArray<*> {
-    val valueColumn = frame.extractValueColumn()
-    val arrayColumn = frame.extractArrayColumn()
-
-    val arraysAreFrames = arrayColumn?.kind() == ColumnKind.Frame
-
-    val data = frame.indices().map { rowIndex ->
-        valueColumn
-            ?.get(rowIndex)
-            ?: arrayColumn?.get(rowIndex)
-                ?.let {
-                    if (arraysAreFrames) encodeFrame(it as AnyFrame) else null
-                }
-            ?: encodeRow(frame, rowIndex)
-    }
-
-    return array(data)
-}
-
 public fun AnyFrame.toJson(prettyPrint: Boolean = false, canonical: Boolean = false): String {
     return json {
         encodeFrame(this@toJson)
     }.toJsonString(prettyPrint, canonical)
 }
 
+/**
+ * Converts the DataFrame to a JSON string representation with additional metadata about serialized data.
+ * It is heavily used to implement some integration features in Kotlin Notebook IntellJ IDEA plugin.
+ *
+ * @param rowLimit The maximum number of top-level dataframe rows to include in the output JSON.
+ * @param nestedRowLimit The maximum number of nested frame rows to include in the output JSON.
+ * If null, all rows are included.
+ * Applied for each frame column recursively
+ * @param prettyPrint Specifies whether the output JSON should be formatted with indentation and line breaks.
+ * @param canonical Specifies whether the output JSON should be in a canonical form.
+ *
+ * @return The DataFrame converted to a JSON string with metadata.
+ */
 public fun AnyFrame.toJsonWithMetadata(
     rowLimit: Int,
-    nestedRowLimit: Int? = null
-): JsonObject {
+    nestedRowLimit: Int? = null,
+    prettyPrint: Boolean = false,
+    canonical: Boolean = false
+): String {
     return json {
-        obj(
-            VERSION to SERIALIZATION_VERSION,
-            METADATA to obj(
-                COLUMNS to columnNames(),
-                NROW to rowsCount(),
-                NCOL to columnsCount()
-            ),
-            KOTLIN_DATAFRAME to encodeFrameWithMetadata(
-                rows().take(rowLimit).toDataFrame(),
-                rowLimit = nestedRowLimit
-            ),
-        )
-    }
+        encodeDataFrameWithMetadata(this@toJsonWithMetadata, rowLimit, nestedRowLimit)
+    }.toJsonString(prettyPrint, canonical)
 }
 
 public fun AnyRow.toJson(prettyPrint: Boolean = false, canonical: Boolean = false): String {
